Adding Fine-Tuned Model Integration
Adding function in utils.py that accesses the fine-tuned model, while leaving access to the base pre-trained model. I also integrated the function into the gui so it can be accessed by a click of a button, there is still the option to use the pretrained model as well. Ontop of this I reorganized the layout of the buttons to fit a more logical arrangement.